Modify VM to be more native compilation friendly
Specifically, there is no way to detect when the end of VM code is
reached, because code is simply dangling off the end of a VMFunction.

The virtual machine relies on there always being an OP_RETURN at the end
of a function if nothing else has happened, but since the native code
generator needs to iterate over all instructions, I've replaced this
with a new OP_RETURN_END instruction which signals the end of code,
whether or not it's ever reached.

Probably should just add a code_size field to VMFunction. In that case
we could also avoid generating OP_RETURN_END if control has already
returned such as (lambda () (if #t #f)).