Add SValue (Stack Value) struct
SValue is derived from Value, but will ultimately contain constructors
and operator= *instead* of Value.

This is part 1 of a refactor intended to make Value into a plain old
data struct. This is part of simplifying native compilation on Windows
and hopefully increasing performance of Windows builds regardless of
whether or not native compilation is enabled.